requirements.txt
attrs==22.2.0
exceptiongroup==1.1.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pytest==7.2.0
tomli==2.0.1
tox
coverage
pytest-cov
pytest-flask
punch.py
wheel
watchdog
flake8
Sphinx
SQLAlchemy
pytest-docker
docker-compose
pyyaml
psycopg2-binary
sqlalchemy_utils